HTML:
No changes needed to the HTML structure.
CSS:

Added a media query to adjust the grid size to 7x7 when the screen width is 600px or less.
JavaScript:

Adjusted the width variable based on the initial screen size.
Modified the createBoard and resetBoard functions to set the width based on the current screen size.
Added an event listener for window resize to adjust the grid size and reset the board if the screen size changes.
With these changes, the grid will change to 7x7 on mobile devices (screen width 600px or less) and revert to 9x9 on larger screens.

The game board will reset when the title is clicked or when the screen size changes.